Put ListenerQueue in separate file
